+++ src/loaders/api/addToTrace.js
@@ -4,6 +4,7 @@
  */
 import { originTime } from '../../common/constants/runtime'
 import { handle } from '../../common/event-emitter/handle'
+import { warn } from '../../common/util/console'
 import { FEATURE_NAMES } from '../features/features'
 import { ADD_TO_TRACE } from './constants'
 import { setupAPI } from './sharedHandlers'
@@ -18,6 +19,11 @@ export function setupAddToTraceAPI (agent) {
       e: (evt.end || evt.start) - originTime,
       o: evt.origin || '',
       t: 'api'
+    }
+
+    if (report.s < 0 || report.e < 0 || report.e < report.s) {
+      warn(61, { start: report.s, end: report.e })
+      return
     }
 
     handle('bstApi', [report], undefined, FEATURE_NAMES.sessionTrace, agent.ee)