+++ tests/specs/soft_navigations/payloads.e2e.js
@@ -1,4 +1,4 @@
-import { notSafari, supportsFirstPaint } from '../../../tools/browser-matcher/common-matchers.mjs'
+import { notIOS, notSafari, supportsFirstPaint } from '../../../tools/browser-matcher/common-matchers.mjs'
 import { testInteractionEventsRequest, testErrorsRequest } from '../../../tools/testing-server/utils/expect-tests'
 
 describe('attribution tests', () => {
@@ -27,11 +27,12 @@ describe('attribution tests', () => {
           await browser.testHandle.assetURL('soft_navigations/action-text.html', configWithDenyList)
         ).then(() => browser.waitForAgentLoad())
       ])
+      const documentReferrer = await browser.execute(() => document.referrer)
       const ipl = interactionHarvests[0].request.body[0]
 
       expect(ipl.trigger).toEqual('initialPageLoad')
-      expect(ipl.children.length).toEqual(0)
       expect(ipl.isRouteChange).not.toBeTruthy()
+      if (browserMatch(notIOS)) expect(ipl.oldURL).toEqual(documentReferrer) // ios on lambdatest appears to return the wrong value for referrer when using browser.execute, which breaks this test condition. Confirmed referrer behavior works in real env
 
       ;[interactionHarvests] = await Promise.all([
         interactionsCapture.waitForResult({ totalCount: 2 }),
@@ -54,11 +55,13 @@ describe('attribution tests', () => {
           await browser.testHandle.assetURL('soft_navigations/action-text.html', configWithDenyList)
         ).then(() => browser.waitForAgentLoad())
       ])
+      const documentReferrer = await browser.execute(() => document.referrer)
       const ipl = interactionHarvests[0].request.body[0]
 
       expect(ipl.trigger).toEqual('initialPageLoad')
       expect(ipl.children.length).toEqual(0)
       expect(ipl.isRouteChange).not.toBeTruthy()
+      if (browserMatch(notIOS)) expect(ipl.oldURL).toEqual(documentReferrer) // ios on lambdatest appears to return the wrong value for referrer when using browser.execute, which breaks this test condition. Confirmed referrer behavior works in real env
 
       ;[interactionHarvests] = await Promise.all([
         interactionsCapture.waitForResult({ totalCount: 2 }),
@@ -81,11 +84,13 @@ describe('attribution tests', () => {
           await browser.testHandle.assetURL('soft_navigations/action-text.html', configWithDenyList)
         ).then(() => browser.waitForAgentLoad())
       ])
+      const documentReferrer = await browser.execute(() => document.referrer)
       const ipl = interactionHarvests[0].request.body[0]
 
       expect(ipl.trigger).toEqual('initialPageLoad')
       expect(ipl.children.length).toEqual(0)
       expect(ipl.isRouteChange).not.toBeTruthy()
+      if (browserMatch(notIOS)) expect(ipl.oldURL).toEqual(documentReferrer) // ios on lambdatest appears to return the wrong value for referrer when using browser.execute, which breaks this test condition. Confirmed referrer behavior works in real env
 
       ;[interactionHarvests] = await Promise.all([
         interactionsCapture.waitForResult({ totalCount: 2 }),
@@ -108,11 +113,13 @@ describe('attribution tests', () => {
           await browser.testHandle.assetURL('soft_navigations/action-text.html', configWithDenyList)
         ).then(() => browser.waitForAgentLoad())
       ])
+      const documentReferrer = await browser.execute(() => document.referrer)
       const ipl = interactionHarvests[0].request.body[0]
 
       expect(ipl.trigger).toEqual('initialPageLoad')
       expect(ipl.children.length).toEqual(0)
       expect(ipl.isRouteChange).not.toBeTruthy()
+      if (browserMatch(notIOS)) expect(ipl.oldURL).toEqual(documentReferrer) // ios on lambdatest appears to return the wrong value for referrer when using browser.execute, which breaks this test condition. Confirmed referrer behavior works in real env
 
       ;[interactionHarvests] = await Promise.all([
         interactionsCapture.waitForResult({ totalCount: 2 }),
@@ -136,10 +143,13 @@ describe('attribution tests', () => {
           .then(() => $('body').click())
       ])
 
+      const documentReferrer = await browser.execute(() => document.referrer)
+
       const [{ request: { body: [ipl] } }, { request: { body: [rc] } }] = interactionHarvests
 
       expect(ipl.trigger).toEqual('initialPageLoad')
       expect(ipl.navTiming).toEqual(expect.any(Object))
+      if (browserMatch(notIOS)) expect(ipl.oldURL).toEqual(documentReferrer) // ios on lambdatest appears to return the wrong value for referrer when using browser.execute, which breaks this test condition. Confirmed referrer behavior works in real env
       if (browserMatch(supportsFirstPaint)) expect(ipl.firstPaint).toBeGreaterThan(0)
       else expect(ipl.firstPaint).toBeNull()
       expect(ipl.firstContentfulPaint).toBeGreaterThan(0)
@@ -148,6 +158,7 @@ describe('attribution tests', () => {
       expect(rc.navTiming).toBeNull()
       expect(rc.firstPaint).toBeNull()
       expect(rc.firstContentfulPaint).toBeNull()
+      expect(rc.oldURL).toEqual(ipl.newURL)
     })
   })